Type usage consistency, braces, etc.

+++ crypto/include/aes_icm.h
@@ -55,7 +55,7 @@ typedef struct {
     v128_t offset;                        /* initial offset value             */
     v128_t keystream_buffer;              /* buffers bytes of keystream       */
     srtp_aes_expanded_key_t expanded_key; /* the cipher key                   */
-    int bytes_in_buffer;                  /* number of unused bytes in buffer */
+    size_t bytes_in_buffer;               /* number of unused bytes in buffer */
     size_t key_size;                      /* AES key size + 14 byte SALT */
 } srtp_aes_icm_ctx_t;
